+++ lua/jokers/heavenlykitsune.lua
@@ -3,7 +3,6 @@ SMODS.Joker {
     key = 'kitsune',
 	pos = { x = 2, y = 0 }, soul_pos = { x = 0, y = 0 },
     set_badges = function(self, card, badges) if (self.discovered) then
-		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Bun)
 		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.SO)
 	end end,
     discovered = true,